requirements.txt
matplotlib==3.1.2
numpy==1.16.6
Pillow==9.1.0
scipy==1.7.3
